Time: 0 ms (100.00%) | Memory: 7.6 MB (41.78%) - LeetSync

+++ 2323-minimum-bit-flips-to-convert-number/minimum-bit-flips-to-convert-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int minBitFlips(int start, int goal) {
         int  a = start ^ goal;
-        long long cnt = 0;
+        int cnt = 0;
         while (a>0) {
             if (a % 2 == 1) 
                 cnt++;
